Add snippets submodule and update installer.
This way, I can have my own snippets for snipMate in the same hierarchy
without having to fork snipMate *or* snipmate-nodejs. Also did some work
to make my vim installer script closer to idempotent.

+++ install.sh
@@ -3,14 +3,29 @@ set -e
 
 cd ~
 
-mv .vimrc .vimrc-
-ln -sf .vim/vimrc .vimrc
-mv .gvimrc .gvimrc-
-ln -sf .vim/gvimrc .gvimrc
+if [ X".vim/vimrc" != `stat -f '%Y' ~/.vimrc` ]
+then
+    mv .vimrc .vimrc-
+    ln -sf .vim/vimrc .vimrc
+fi
+
+if [ X".vim/gvimrc" != `stat -f '%Y' ~/.gvimrc` ]
+then
+    mv .gvimrc .gvimrc-
+    ln -sf .vim/gvimrc .gvimrc
+fi
 
 cd ~/.vim
 
 # make sure that submodules are correctly configured
 git submodule init
-git submodule update
 git submodule sync
+git submodule update
+
+# set up snippets
+mkdir -p snippets/javascript
+cd snippets/javascript
+for snippet in ../../resources/snipmate-nodejs/snippets/javascript/*.snippet
+do
+    ln -s $snippet .
+done
